+++ inboxen/assets.py
@@ -37,8 +37,8 @@ css = Bundle(
 
 js = Bundle(
     "thirdparty/jquery/dist/jquery.js",
+    "js/utils.js",
     "js/alert.js",
-    "js/menu.js",
     "js/home.js",
     "js/search.js",
     "js/inbox.js",